# docs/infrastructure/api_v2/current-openapi-spec.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Assixx API Documentation",
    "version": "1.0.0",
    "description": "Multi-Tenant SaaS Platform für Industrieunternehmen",
    "contact": { "name": "SCS-Technik", "email": "support@scs-technik.de" },
    "license": { "name": "Proprietary", "url": "https://assixx.com/license" }
  },
  "servers": [
    { "url": "https://api.assixx.com/api", "description": "Production Server" }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter your JWT token in the format: Bearer <token>"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": { "type": "string", "description": "Error message" },
          "message": {
            "type": "string",
            "description": "Error message (alternative field)"
          },
          "code": { "type": "string", "description": "Error code" },
          "details": {
            "type": "object",
            "description": "Additional error details"
          }
        }
      },
...
      "User": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "description": "User ID" },
          "username": { "type": "string", "description": "Username" },
          "email": {
            "type": "string",
            "format": "email",
            "description": "Email address"
          },
          "first_name": { "type": "string", "description": "First name" },
          "last_name": { "type": "string", "description": "Last name" },
          "role": {
            "type": "string",
            "enum": ["root", "admin", "employee"],
            "description": "User role"
          },
          "tenant_id": { "type": "integer", "description": "Tenant ID" },
          "department_id": {
            "type": "integer",
            "nullable": true,
            "description": "Department ID"
          },
          "is_active": {
            "type": "boolean",
            "description": "Whether the user is active"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "description": "Creation timestamp"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "description": "Last update timestamp"
          }
        },
        "required": ["id", "username", "email", "role", "tenant_id"]
      },
...
      "Department": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "description": "Department ID" },
          "name": { "type": "string", "description": "Department name" },
          "description": {
            "type": "string",
            "nullable": true,
            "description": "Department description"
          },
          "manager_id": {
            "type": "integer",
            "nullable": true,
            "description": "User ID of department manager"
          },
          "manager_name": {
            "type": "string",
            "nullable": true,
            "description": "Manager display name"
          },
          "parent_id": {
            "type": "integer",
            "nullable": true,
            "description": "Parent department ID"
          },
          "parent_name": {
            "type": "string",
            "nullable": true,
            "description": "Parent department name"
          },
          "location": {
            "type": "string",
            "nullable": true,
            "description": "Physical location"
          },
          "cost_center": {
            "type": "string",
            "nullable": true,
            "description": "Cost center code"
          },
          "employee_count": {
            "type": "integer",
            "description": "Number of employees in department"
          },
          "tenant_id": { "type": "integer", "description": "Tenant ID" },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "description": "Creation timestamp"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "description": "Last update timestamp"
          }
        },
        "required": ["id", "name", "tenant_id"]
      },
...
    }
  },
  "security": [{ "bearerAuth": [] }],
...
  "paths": {
...
    "/departments": {
      "post": {
        "summary": "Create a new department",
        "description": "Create a new department (Admin/Root only)",
        "tags": ["Departments"],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name"],
                "properties": {
                  "name": {
                    "type": "string",
                    "description": "Department name",
                    "example": "Produktion"
                  },
                  "description": {
                    "type": "string",
                    "description": "Department description",
                    "example": "Produktionsabteilung für alle Fertigungsprozesse"
                  },
                  "manager_id": {
                    "type": "integer",
                    "description": "User ID of department manager",
                    "example": 42
                  },
                  "parent_id": {
                    "type": "integer",
                    "description": "Parent department ID for hierarchical structure",
                    "example": 1
                  },
                  "location": {
                    "type": "string",
                    "description": "Physical location",
                    "example": "Gebäude A, 2. Stock"
                  },
                  "cost_center": {
                    "type": "string",
                    "description": "Cost center code",
                    "example": "CC-PROD-001"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Department created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Abteilung erfolgreich erstellt"
                    },
                    "departmentId": { "type": "integer", "example": 5 }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Abteilungsname ist erforderlich"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          },
          "403": {
            "description": "Forbidden - Not admin/root",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Zugriff verweigert"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      },
      "get": {
        "summary": "Get all departments",
        "description": "Retrieve all departments for the tenant",
        "tags": ["Departments"],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Departments retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Department" }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      }
    },
    "/departments/{id}": {
      "get": {
        "summary": "Get department by ID",
        "description": "Retrieve a specific department by its ID",
        "tags": ["Departments"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": { "type": "integer" },
            "description": "Department ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Department retrieved successfully",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Department" }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          },
          "404": {
            "description": "Department not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Abteilung nicht gefunden"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Update department",
        "description": "Update an existing department (Admin/Root only)",
        "tags": ["Departments"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": { "type": "integer" },
            "description": "Department ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "description": "Department name",
                    "example": "Produktion"
                  },
                  "description": {
                    "type": "string",
                    "description": "Department description"
                  },
                  "manager_id": {
                    "type": "integer",
                    "description": "User ID of department manager"
                  },
                  "parent_id": {
                    "type": "integer",
                    "description": "Parent department ID"
                  },
                  "location": {
                    "type": "string",
                    "description": "Physical location"
                  },
                  "cost_center": {
                    "type": "string",
                    "description": "Cost center code"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Department updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Abteilung erfolgreich aktualisiert"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Eine Abteilung kann nicht sich selbst als Übergeordnete haben"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          },
          "403": {
            "description": "Forbidden - Not admin/root",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Zugriff verweigert"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Department not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Abteilung nicht gefunden"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete department",
        "description": "Delete a department (Admin/Root only). Cannot delete if users are assigned.",
        "tags": ["Departments"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": { "type": "integer" },
            "description": "Department ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Department deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Abteilung erfolgreich gelöscht"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Cannot delete - has assigned users",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Diese Abteilung kann nicht gelöscht werden, da ihr noch Benutzer zugeordnet sind"
                    },
                    "users": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/User" }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          },
          "403": {
            "description": "Forbidden - Not admin/root",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Zugriff verweigert"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Department not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Abteilung nicht gefunden"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      }
    },
    "/departments/{id}/members": {
      "get": {
        "summary": "Get department members",
        "description": "Retrieve all users assigned to a specific department",
        "tags": ["Departments"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": { "type": "integer" },
            "description": "Department ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Department members retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": { "type": "integer" },
                      "username": { "type": "string" },
                      "email": { "type": "string" },
                      "first_name": { "type": "string" },
                      "last_name": { "type": "string" },
                      "role": {
                        "type": "string",
                        "enum": ["root", "admin", "employee"]
                      },
                      "position": { "type": "string" },
                      "is_active": { "type": "boolean" }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          },
          "404": {
            "description": "Department not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Abteilung nicht gefunden"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      }
    },
...
  }
}